requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bidict==0.22.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.1.2
colorama==0.4.4
cryptography==36.0.1
cssselect2==0.6.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-import-export==2.7.1
django-oauth-toolkit==1.7.0
django-phonenumber-field==6.1.0
django-tz-detect==0.4.0
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
fonttools==4.33.3
future==0.18.2
html5lib==1.1
idna==3.3
jeepney==0.7.1
Jinja2==3.1.1
jwcrypto==1.0
lml==0.1.0
loguru==0.5.3
MarkupPy==1.14
MarkupSafe==2.1.1
notify-py==0.3.3
numpy==1.22.2
oauthlib==3.2.0
pdfkit==1.0.0
phonenumbers==8.12.43
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
#pycrypto==2.6.1
pycryptodome==3.15.0
PyJWT==2.3.0
PyPDF2==1.26.0
python-barcode==0.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
requests==2.27.1
six==1.16.0
speaklater==1.3
sqlparse==0.4.2
tablib==3.2.0
tinycss2==1.1.1
twilio==7.10.0
tzdata==2021.5
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==2.1.1
win32-setctime==1.1.0
wrapt==1.13.3
WTForms==3.0.1
xhtml2pdf==0.2.5
xlwt==1.3.0

